refactor: share dismiss logic for annotation toolbars
